refactor: modularize connection handling
- move common functions to `common.rs`
- create `linux.rs` for Linux-specific connection handling
- create `macos.rs` for macOS-specific connection handling
- update `mod.rs` to include new modules and functions
